[Fix] 不足しているキャストに追加対応．

+++ src/wizard/wizard-special-process.cpp
@@ -503,7 +503,7 @@ void wiz_reset_race(player_type *creature_ptr)
     sprintf(ppp, "Race (0-%d): ", MAX_RACES - 1);
 
     char tmp_val[160];
-    sprintf(tmp_val, "%d", creature_ptr->prace);
+    sprintf(tmp_val, "%d", static_cast<int>(creature_ptr->prace));
 
     if (!get_string(ppp, tmp_val, 2))
         return;
